+++ Elements/src/Geometry/Profiles/HSSPipeProfileType.cs
@@ -1,0 +1,148 @@
+#pragma warning disable CS1591
+namespace Elements.Geometry.Profiles
+{
+    public enum HSSPipeProfileType
+    {
+        HSS20_000x0_500,
+        HSS20_000x0_375,
+        HSS18_000x0_500,
+        HSS18_000x0_375,
+        HSS16_000x0_500,
+        HSS16_000x0_438,
+        HSS16_000x0_375,
+        HSS16_000x0_312,
+        HSS14_000x0_500,
+        HSS14_000x0_375,
+        HSS14_000x0_312,
+        HSS12_750x0_500,
+        HSS12_750x0_375,
+        HSS12_750x0_250,
+        HSS12_500x0_625,
+        HSS12_500x0_500,
+        HSS12_500x0_375,
+        HSS12_500x0_312,
+        HSS12_500x0_250,
+        HSS12_500x0_188,
+        HSS11_250x0_625,
+        HSS11_250x0_500,
+        HSS11_250x0_375,
+        HSS11_250x0_312,
+        HSS11_250x0_250,
+        HSS11_250x0_188,
+        HSS10_750x0_500,
+        HSS10_750x0_250,
+        HSS10_000x0_625,
+        HSS10_000x0_500,
+        HSS10_000x0_375,
+        HSS10_000x0_312,
+        HSS10_000x0_250,
+        HSS10_000x0_188,
+        HSS9_625x0_500,
+        HSS9_625x0_375,
+        HSS9_625x0_312,
+        HSS9_625x0_250,
+        HSS9_625x0_188,
+        HSS8_750x0_500,
+        HSS8_750x0_375,
+        HSS8_750x0_312,
+        HSS8_750x0_250,
+        HSS8_750x0_188,
+        HSS8_625x0_500,
+        HSS8_625x0_375,
+        HSS8_625x0_322,
+        HSS8_625x0_250,
+        HSS8_625x0_188,
+        HSS7_625x0_125,
+        HSS7_500x0_500,
+        HSS7_500x0_375,
+        HSS7_500x0_312,
+        HSS7_500x0_250,
+        HSS7_500x0_188,
+        HSS7_000x0_500,
+        HSS7_000x0_375,
+        HSS7_000x0_312,
+        HSS7_000x0_250,
+        HSS7_000x0_188,
+        HSS7_000x0_125,
+        HSS6_875x0_500,
+        HSS6_875x0_375,
+        HSS6_875x0_312,
+        HSS6_875x0_250,
+        HSS6_875x0_188,
+        HSS6_625x0_500,
+        HSS6_625x0_432,
+        HSS6_625x0_375,
+        HSS6_625x0_312,
+        HSS6_625x0_280,
+        HSS6_625x0_250,
+        HSS6_625x0_188,
+        HSS6_625x0_125,
+        HSS6_125x0_500,
+        HSS6_125x0_375,
+        HSS6_125x0_312,
+        HSS6_125x0_250,
+        HSS6_125x0_188,
+        HSS6_000x0_500,
+        HSS6_000x0_375,
+        HSS6_000x0_312,
+        HSS6_000x0_280,
+        HSS6_000x0_250,
+        HSS6_000x0_188,
+        HSS6_000x0_125,
+        HSS5_563x0_375,
+        HSS5_563x0_258,
+        HSS5_563x0_188,
+        HSS5_563x0_134,
+        HSS5_500x0_500,
+        HSS5_500x0_375,
+        HSS5_500x0_258,
+        HSS5_000x0_500,
+        HSS5_000x0_375,
+        HSS5_000x0_312,
+        HSS5_000x0_258,
+        HSS5_000x0_250,
+        HSS5_000x0_188,
+        HSS5_000x0_125,
+        HSS4_500x0_337,
+        HSS4_500x0_237,
+        HSS4_500x0_188,
+        HSS4_500x0_125,
+        HSS4_000x0_337,
+        HSS4_000x0_313,
+        HSS4_000x0_250,
+        HSS4_000x0_237,
+        HSS4_000x0_226,
+        HSS4_000x0_220,
+        HSS4_000x0_188,
+        HSS4_000x0_125,
+        HSS3_500x0_313,
+        HSS3_500x0_300,
+        HSS3_500x0_250,
+        HSS3_500x0_216,
+        HSS3_500x0_203,
+        HSS3_500x0_188,
+        HSS3_500x0_125,
+        HSS3_000x0_300,
+        HSS3_000x0_250,
+        HSS3_000x0_216,
+        HSS3_000x0_203,
+        HSS3_000x0_188,
+        HSS3_000x0_152,
+        HSS3_000x0_134,
+        HSS3_000x0_120,
+        HSS2_875x0_250,
+        HSS2_875x0_203,
+        HSS2_875x0_188,
+        HSS2_875x0_125,
+        HSS2_500x0_250,
+        HSS2_500x0_188,
+        HSS2_500x0_125,
+        HSS2_375x0_250,
+        HSS2_375x0_218,
+        HSS2_375x0_188,
+        HSS2_375x0_154,
+        HSS2_375x0_125,
+        HSS1_900x0_145,
+        HSS1_660x0_140
+    }
+}
