+++ Elements/src/Geometry/Profiles/LProfileType.cs
@@ -1,0 +1,144 @@
+#pragma warning disable CS1591
+namespace Elements.Geometry.Profiles
+{
+    public enum LProfileType
+    {
+        L12X12X1__3_8,
+        L12X12X1__1_4,
+        L12X12X1__1_8,
+        L12X12X1,
+        L10X10X1__3_8,
+        L10X10X1__1_4,
+        L10X10X1__1_8,
+        L10X10X1,
+        L10X10X7_8,
+        L10X10X3_4,
+        L8X8X1__1_8,
+        L8X8X1,
+        L8X8X7_8,
+        L8X8X3_4,
+        L8X8X5_8,
+        L8X8X9_16,
+        L8X8X1_2,
+        L8X6X1,
+        L8X6X7_8,
+        L8X6X3_4,
+        L8X6X5_8,
+        L8X6X9_16,
+        L8X6X1_2,
+        L8X6X7_16,
+        L8X4X1,
+        L8X4X7_8,
+        L8X4X3_4,
+        L8X4X5_8,
+        L8X4X9_16,
+        L8X4X1_2,
+        L8X4X7_16,
+        L7X4X3_4,
+        L7X4X5_8,
+        L7X4X1_2,
+        L7X4X7_16,
+        L7X4X3_8,
+        L6X6X1,
+        L6X6X7_8,
+        L6X6X3_4,
+        L6X6X5_8,
+        L6X6X9_16,
+        L6X6X1_2,
+        L6X6X7_16,
+        L6X6X3_8,
+        L6X6X5_16,
+        L6X4X7_8,
+        L6X4X3_4,
+        L6X4X5_8,
+        L6X4X9_16,
+        L6X4X1_2,
+        L6X4X7_16,
+        L6X4X3_8,
+        L6X4X5_16,
+        L6X3__1_2X1_2,
+        L6X3__1_2X3_8,
+        L6X3__1_2X5_16,
+        L5X5X7_8,
+        L5X5X3_4,
+        L5X5X5_8,
+        L5X5X1_2,
+        L5X5X7_16,
+        L5X5X3_8,
+        L5X5X5_16,
+        L5X3__1_2X3_4,
+        L5X3__1_2X5_8,
+        L5X3__1_2X1_2,
+        L5X3__1_2X3_8,
+        L5X3__1_2X5_16,
+        L5X3__1_2X1_4,
+        L5X3X1_2,
+        L5X3X7_16,
+        L5X3X3_8,
+        L5X3X5_16,
+        L5X3X1_4,
+        L4X4X3_4,
+        L4X4X5_8,
+        L4X4X1_2,
+        L4X4X7_16,
+        L4X4X3_8,
+        L4X4X5_16,
+        L4X4X1_4,
+        L4X3__1_2X1_2,
+        L4X3__1_2X3_8,
+        L4X3__1_2X5_16,
+        L4X3__1_2X1_4,
+        L4X3X5_8,
+        L4X3X1_2,
+        L4X3X3_8,
+        L4X3X5_16,
+        L4X3X1_4,
+        L3__1_2X3__1_2X1_2,
+        L3__1_2X3__1_2X7_16,
+        L3__1_2X3__1_2X3_8,
+        L3__1_2X3__1_2X5_16,
+        L3__1_2X3__1_2X1_4,
+        L3__1_2X3X1_2,
+        L3__1_2X3X7_16,
+        L3__1_2X3X3_8,
+        L3__1_2X3X5_16,
+        L3__1_2X3X1_4,
+        L3__1_2X2__1_2X1_2,
+        L3__1_2X2__1_2X3_8,
+        L3__1_2X2__1_2X5_16,
+        L3__1_2X2__1_2X1_4,
+        L3X3X1_2,
+        L3X3X7_16,
+        L3X3X3_8,
+        L3X3X5_16,
+        L3X3X1_4,
+        L3X3X3_16,
+        L3X2__1_2X1_2,
+        L3X2__1_2X7_16,
+        L3X2__1_2X3_8,
+        L3X2__1_2X5_16,
+        L3X2__1_2X1_4,
+        L3X2__1_2X3_16,
+        L3X2X1_2,
+        L3X2X3_8,
+        L3X2X5_16,
+        L3X2X1_4,
+        L3X2X3_16,
+        L2__1_2X2__1_2X1_2,
+        L2__1_2X2__1_2X3_8,
+        L2__1_2X2__1_2X5_16,
+        L2__1_2X2__1_2X1_4,
+        L2__1_2X2__1_2X3_16,
+        L2__1_2X2X3_8,
+        L2__1_2X2X5_16,
+        L2__1_2X2X1_4,
+        L2__1_2X2X3_16,
+        L2__1_2X1__1_2X1_4,
+        L2__1_2X1__1_2X3_16,
+        L2X2X3_8,
+        L2X2X5_16,
+        L2X2X1_4,
+        L2X2X3_16,
+        L2X2X1_8,
+    }
+}
