+++ Elements/src/Geometry/Profiles/RHSProfileFactory.cs
@@ -3,256 +3,13 @@ using System;
 
 namespace Elements.Geometry.Profiles
 {
-    public enum RHSProfileType
-    {
-        RHS60x30x1_6,
-        RHS60x30x2,
-        RHS60x30x2_3,
-        RHS60x30x2_5,
-        RHS60x30x3,
-        RHS60x30x3_2,
-        RHS60x30x4,
-        RHS60x30x4_5,
-        RHS60x40x1_6,
-        RHS60x40x2,
-        RHS60x40x2_3,
-        RHS60x40x2_5,
-        RHS60x40x3,
-        RHS60x40x3_2,
-        RHS60x40x4,
-        RHS60x40x4_5,
-        RHS70x40x1_6,
-        RHS70x40x2,
-        RHS70x40x2_3,
-        RHS70x40x2_5,
-        RHS70x40x3,
-        RHS70x40x3_2,
-        RHS70x40x4,
-        RHS70x40x4_5,
-        RHS75x45x1_6,
-        RHS75x45x2,
-        RHS75x45x2_3,
-        RHS75x45x2_5,
-        RHS75x45x3,
-        RHS75x45x3_2,
-        RHS75x45x4,
-        RHS75x45x4_5,
-        RHS75x45x5,
-        RHS75x45x6,
-        RHS75x50x2_3,
-        RHS75x50x2_5,
-        RHS75x50x3,
-        RHS75x50x3_2,
-        RHS75x50x4,
-        RHS75x50x4_5,
-        RHS75x50x5,
-        RHS75x50x6,
-        RHS100x40x1_6,
-        RHS100x40x2,
-        RHS100x40x2_3,
-        RHS100x40x2_5,
-        RHS100x40x3,
-        RHS100x40x3_2,
-        RHS100x40x4,
-        RHS100x40x4_5,
-        RHS100x50x1_6,
-        RHS100x50x2,
-        RHS100x50x2_3,
-        RHS100x50x2_5,
-        RHS100x50x3,
-        RHS100x50x3_2,
-        RHS100x50x4,
-        RHS100x50x4_5,
-        RHS100x50x5,
-        RHS100x50x6,
-        RHS125x40x1_6,
-        RHS125x40x2,
-        RHS125x40x2_3,
-        RHS125x40x2_5,
-        RHS125x40x3,
-        RHS125x40x3_2,
-        RHS125x40x4,
-        RHS125x40x4_5,
-        RHS125x50x1_6,
-        RHS125x50x2,
-        RHS125x50x2_3,
-        RHS125x50x2_5,
-        RHS125x50x3,
-        RHS125x50x3_2,
-        RHS125x50x4,
-        RHS125x50x4_5,
-        RHS125x50x5,
-        RHS125x50x6,
-        RHS125x75x1_6,
-        RHS125x75x2,
-        RHS125x75x2_3,
-        RHS125x75x2_5,
-        RHS125x75x3,
-        RHS125x75x3_2,
-        RHS125x75x4,
-        RHS125x75x4_5,
-        RHS125x75x5,
-        RHS125x75x6,
-        RHS125x75x6_3,
-        RHS125x75x8,
-        RHS125x75x9,
-        RHS150x50x2_3,
-        RHS150x50x2_5,
-        RHS150x50x3,
-        RHS150x50x3_2,
-        RHS150x50x4,
-        RHS150x50x4_5,
-        RHS150x50x5,
-        RHS150x50x6,
-        RHS150x75x3,
-        RHS150x75x3_2,
-        RHS150x75x4,
-        RHS150x75x4_5,
-        RHS150x75x5,
-        RHS150x75x6,
-        RHS150x75x6_3,
-        RHS150x75x8,
-        RHS150x75x9,
-        RHS150x80x3,
-        RHS150x80x3_2,
-        RHS150x80x4,
-        RHS150x80x4_5,
-        RHS150x80x5,
-        RHS150x80x6,
-        RHS150x100x2_3,
-        RHS150x100x2_5,
-        RHS150x100x3,
-        RHS150x100x3_2,
-        RHS150x100x4,
-        RHS150x100x4_5,
-        RHS150x100x5,
-        RHS150x100x6,
-        RHS150x100x6_3,
-        RHS150x100x8,
-        RHS150x100x9,
-        RHS150x100x10,
-        RHS150x100x12,
-        RHS175x125x4_5,
-        RHS175x125x5,
-        RHS175x125x6,
-        RHS175x125x6_3,
-        RHS175x125x8,
-        RHS175x125x9,
-        RHS175x125x10,
-        RHS175x125x12,
-        RHS200x100x3_2,
-        RHS200x100x4,
-        RHS200x100x4_5,
-        RHS200x100x5,
-        RHS200x100x6,
-        RHS200x100x6_3,
-        RHS200x100x8,
-        RHS200x100x9,
-        RHS200x100x10,
-        RHS200x100x12,
-        RHS200x100x12_5,
-        RHS200x150x4_5,
-        RHS200x150x5,
-        RHS200x150x6,
-        RHS200x150x6_3,
-        RHS200x150x8,
-        RHS200x150x9,
-        RHS200x150x10,
-        RHS200x150x12,
-        RHS200x150x12_5,
-        RHS250x100x4_5,
-        RHS250x100x5,
-        RHS250x100x6,
-        RHS250x100x6_3,
-        RHS250x100x8,
-        RHS250x100x9,
-        RHS250x100x10,
-        RHS250x100x12,
-        RHS225x175x4_5,
-        RHS225x175x5,
-        RHS225x175x6,
-        RHS250x150x4_5,
-        RHS250x150x5,
-        RHS250x150x6,
-        RHS250x150x6_3,
-        RHS250x150x8,
-        RHS250x150x9,
-        RHS250x150x10,
-        RHS250x150x12,
-        RHS250x150x12_5,
-        RHS300x100x4_5,
-        RHS300x100x5,
-        RHS300x100x6,
-        RHS300x100x6_3,
-        RHS300x100x8,
-        RHS300x100x9,
-        RHS250x200x6,
-        RHS250x200x6_3,
-        RHS250x200x8,
-        RHS250x200x9,
-        RHS250x200x10,
-        RHS250x200x12,
-        RHS250x200x12_5,
-        RHS300x150x6,
-        RHS300x150x6_3,
-        RHS300x150x8,
-        RHS300x150x9,
-        RHS300x150x10,
-        RHS300x150x12,
-        RHS300x150x12_5,
-        RHS300x200x6,
-        RHS300x200x6_3,
-        RHS300x200x8,
-        RHS300x200x9,
-        RHS300x200x10,
-        RHS300x200x12,
-        RHS300x200x12_5,
-        RHS300x200x16,
-        RHS350x150x6,
-        RHS350x150x6_3,
-        RHS350x150x8,
-        RHS350x150x9,
-        RHS350x150x10,
-        RHS350x150x12,
-        RHS350x150x12_5,
-        RHS350x250x6,
-        RHS350x250x6_3,
-        RHS350x250x8,
-        RHS350x250x9,
-        RHS350x250x10,
-        RHS350x250x12,
-        RHS350x250x12_5,
-        RHS350x250x16,
-        RHS400x200x6,
-        RHS400x200x6_3,
-        RHS400x200x8,
-        RHS400x200x9,
-        RHS400x200x10,
-        RHS400x200x12,
-        RHS400x200x12_5,
-        RHS400x300x9,
-        RHS400x300x10,
-        RHS400x300x12,
-        RHS400x300x12_5,
-        RHS400x300x16,
-        RHS500x300x12,
-        RHS500x300x12_5,
-        RHS500x300x16,
-        RHS500x300x19,
-        RHS600x300x12,
-        RHS600x300x12_5,
-        RHS600x300x16,
-        RHS600x300x19,
-        RHS600x300x22
-    }
-
     /// <summary>
     /// A profile factory for rectangular hollow steel profiles.
     /// </summary>
-    public sealed class RHSProfileFactory : ProfileFactory<RHSProfileType, RHSProfile>
+    public sealed class RHSProfileFactory : ProfileFactoryBase<RHSProfileType, RHSProfile>
     {
         /// Nominal Size,Nominal Size,Thickness,Mass,Area,Second Moment of Area,Second Moment of Area,Radius of Gyration,Radius of Gyration,Elastic Section Modulus,Elastic Section Modulus,Plastic Section modulus,Plastic Section modulus,Torsional Inertia,Torsional Modulus,superficial area per,Nominal Length
-        private static string _data = @"A,B,t,M,Ix,Iy,ix,iy,Zx,Zy,Zpx,Zpy,ti,tm,sa,nl,J
+        private static readonly string _data = @"A,B,t,M,Ix,Iy,ix,iy,Zx,Zy,Zpx,Zpy,ti,tm,sa,nl,J
 60,30,1.6,2.13,2.712,12.5,4.25,2.15,1.25,4.16,2.83,5.19,3.2,10.3,4.9,0.175,470
 60,30,2,2.62,3.337,15.1,5.08,2.12,1.23,5.02,3.39,6.31,3.89,12.6,5.88,0.173,382
 60,30,2.3,2.98,3.792,16.8,5.65,2.11,1.22,5.61,3.76,7.11,4.37,14.1,6.56,0.172,336
