+++ Elements/src/Geometry/Profiles/SHSProfileType.cs
@@ -1,0 +1,176 @@
+#pragma warning disable CS1591
+namespace Elements.Geometry.Profiles
+{
+    public enum SHSProfileType
+    {
+        SHS50xs00x1_6,
+        SHS50xs00x2,
+        SHS50xs00x2_3,
+        SHS50xs00x2_5,
+        SHS50xs00x3,
+        SHS50xs00x3_2,
+        SHS50xs00x4,
+        SHS50xs00x4_5,
+        SHS50xs00x5,
+        SHS50xs00x6,
+        SHS60xs00x1_6,
+        SHS60xs00x2,
+        SHS60xs00x2_3,
+        SHS60xs00x2_5,
+        SHS60xs00x3,
+        SHS60xs00x3_2,
+        SHS60xs00x4,
+        SHS60xs00x4_5,
+        SHS60xs00x5,
+        SHS60xs00x6,
+        SHS75xs00x1_6,
+        SHS75xs00x2,
+        SHS75xs00x2_3,
+        SHS75xs00x2_5,
+        SHS75xs00x3,
+        SHS75xs00x3_2,
+        SHS75xs00x4,
+        SHS75xs00x4_5,
+        SHS75xs00x5,
+        SHS75xs00x6,
+        SHS75xs00x6_3,
+        SHS75xs00x8,
+        SHS80xs00x1_6,
+        SHS80xs00x2,
+        SHS80xs00x2_3,
+        SHS80xs00x2_5,
+        SHS80xs00x3,
+        SHS80xs00x3_2,
+        SHS80xs00x4,
+        SHS80xs00x4_5,
+        SHS80xs00x5,
+        SHS80xs00x6,
+        SHS80xs00x6_3,
+        SHS80xs00x8,
+        SHS90xs00x2_3,
+        SHS90xs00x2_5,
+        SHS90xs00x3,
+        SHS90xs00x3_2,
+        SHS90xs00x4,
+        SHS90xs00x4_5,
+        SHS90xs00x5,
+        SHS90xs00x6,
+        SHS90xs00x6_3,
+        SHS100xs00x1_6,
+        SHS100xs00x2,
+        SHS100xs00x2_3,
+        SHS100xs00x2_5,
+        SHS100xs00x3,
+        SHS100xs00x3_2,
+        SHS100xs00x4,
+        SHS100xs00x4_5,
+        SHS100xs00x5,
+        SHS100xs00x6,
+        SHS100xs00x6_3,
+        SHS100xs00x8,
+        SHS100xs00x9,
+        SHS100xs00x10,
+        SHS100xs00x12,
+        SHS120xs00x6,
+        SHS120xs00x6_3,
+        SHS120xs00x8,
+        SHS120xs00x9,
+        SHS120xs00x10,
+        SHS120xs00x12,
+        SHS125xs00x2_3,
+        SHS125xs00x2_5,
+        SHS125xs00x3,
+        SHS125xs00x3_2,
+        SHS125xs00x4,
+        SHS125xs00x4_5,
+        SHS125xs00x5,
+        SHS125xs00x6,
+        SHS125xs00x6_3,
+        SHS125xs00x8,
+        SHS125xs00x9,
+        SHS125xs00x10,
+        SHS125xs00x12,
+        SHS125xs00x12_5,
+        SHS150xs00x3_2,
+        SHS150xs00x4,
+        SHS150xs00x4_5,
+        SHS150xs00x5,
+        SHS150xs00x6,
+        SHS150xs00x6_3,
+        SHS150xs00x8,
+        SHS150xs00x9,
+        SHS150xs00x10,
+        SHS150xs00x12,
+        SHS150xs00x12_5,
+        SHS175xs00x4_5,
+        SHS175xs00x5,
+        SHS175xs00x6,
+        SHS175xs00x6_3,
+        SHS175xs00x8,
+        SHS175xs00x9,
+        SHS175xs00x10,
+        SHS175xs00x12,
+        SHS175xs00x12_5,
+        SHS200xs00x4_5,
+        SHS200xs00x5,
+        SHS200xs00x6,
+        SHS200xs00x6_3,
+        SHS200xs00x8,
+        SHS200xs00x9,
+        SHS200xs00x10,
+        SHS200xs00x12,
+        SHS200xs00x12_5,
+        SHS200xs00x16,
+        SHS250xs00x6,
+        SHS250xs00x6_3,
+        SHS250xs00x8,
+        SHS250xs00x9,
+        SHS250xs00x10,
+        SHS250xs00x12,
+        SHS250xs00x12_5,
+        SHS250xs00x16,
+        SHS250xs00x19,
+        SHS300xs00x6,
+        SHS300xs00x6_3,
+        SHS300xs00x8,
+        SHS300xs00x9,
+        SHS300xs00x10,
+        SHS300xs00x12,
+        SHS300xs00x12_5,
+        SHS300xs00x16,
+        SHS300xs00x19,
+        SHS350xs00x9,
+        SHS350xs00x10,
+        SHS350xs00x12,
+        SHS350xs00x12_5,
+        SHS350xs00x16,
+        SHS350xs00x19,
+        SHS350xs00x22,
+        SHS400xs00x9,
+        SHS400xs00x10,
+        SHS400xs00x12,
+        SHS400xs00x12_5,
+        SHS400xs00x16,
+        SHS400xs00x19,
+        SHS400xs00x22,
+        SHS450xs00x9,
+        SHS450xs00x10,
+        SHS450xs00x12,
+        SHS450xs00x12_5,
+        SHS450xs00x16,
+        SHS450xs00x19,
+        SHS450xs00x22,
+        SHS500xs00x9,
+        SHS500xs00x10,
+        SHS500xs00x12,
+        SHS500xs00x12_5,
+        SHS500xs00x16,
+        SHS500xs00x19,
+        SHS500xs00x22,
+        SHS550xs00x12,
+        SHS550xs00x12_5,
+        SHS550xs00x16,
+        SHS550xs00x19,
+        SHS550xs00x22
+    }
+}
