+++ Elements/src/Geometry/Profiles/WideFlangeProfileType.cs
@@ -1,0 +1,323 @@
+#pragma warning disable CS1591
+namespace Elements.Geometry
+{
+    public enum WideFlangeProfileType
+    {
+        W44x335,
+        W44x290,
+        W44x285,
+        W44x262,
+        W44x248,
+        W44x230,
+        W44X224,
+        W44x198,
+        W40x655,
+        W40x593,
+        W40x531,
+        W40x503,
+        W40x480,
+        W40x466,
+        W40x436,
+        W40x431,
+        W40x397,
+        W40x392,
+        W40x372,
+        W40x362,
+        W40x331,
+        W40x328,
+        W40x327,
+        W40x324,
+        W40x321,
+        W40x298,
+        W40x297,
+        W40x278,
+        W40x277,
+        W40x268,
+        W40x264,
+        W40x249,
+        W40x244,
+        W40x235,
+        W40x221,
+        W40x215,
+        W40x211,
+        W40x199,
+        W40x192,
+        W40x183,
+        W40x174,
+        W40x167,
+        W40x149,
+        W36x848,
+        W36x798,
+        W36x720,
+        W36x650,
+        W36x588,
+        W36x527,
+        W36x485,
+        W36x439,
+        W36x393,
+        W36x359,
+        W36x328,
+        W36x300,
+        W36x280,
+        W36x260,
+        W36x256,
+        W36x245,
+        W36x232,
+        W36x230,
+        W36x210,
+        W36x194,
+        W36x182,
+        W36x170,
+        W36x160,
+        W36x150,
+        W36x135,
+        W33x619,
+        W33x567,
+        W33x515,
+        W33x468,
+        W33x424,
+        W33x387,
+        W33x354,
+        W33x318,
+        W33x291,
+        W33x263,
+        W33x241,
+        W33x221,
+        W33x201,
+        W33x169,
+        W33x152,
+        W33x141,
+        W33x130,
+        W33x118,
+        W30x581,
+        W30x526,
+        W30x477,
+        W30x433,
+        W30x391,
+        W30x357,
+        W30x326,
+        W30x292,
+        W30x261,
+        W30x235,
+        W30x211,
+        W30x191,
+        W30x173,
+        W30x148,
+        W30x132,
+        W30x124,
+        W30x116,
+        W30x108,
+        W30x99,
+        W30x90,
+        W27x539,
+        W27x494,
+        W27x448,
+        W27x407,
+        W27x368,
+        W27x336,
+        W27x307,
+        W27x281,
+        W27x258,
+        W27x235,
+        W27x217,
+        W27x194,
+        W27x178,
+        W27x161,
+        W27x146,
+        W27x129,
+        W27x114,
+        W27x102,
+        W27x94,
+        W27x84,
+        W24x492,
+        W24x450,
+        W24x408,
+        W24x370,
+        W24x335,
+        W24x306,
+        W24x279,
+        W24x250,
+        W24x229,
+        W24x207,
+        W24x192,
+        W24x176,
+        W24x162,
+        W24x146,
+        W24x131,
+        W24x117,
+        W24x104,
+        W24x103,
+        W24x94,
+        W24x84,
+        W24x76,
+        W24x68,
+        W24x62,
+        W24x55,
+        W21x402,
+        W21x364,
+        W21x333,
+        W21x300,
+        W21x275,
+        W21x248,
+        W21x223,
+        W21x201,
+        W21x182,
+        W21x166,
+        W21x147,
+        W21x132,
+        W21x122,
+        W21x111,
+        W21x101,
+        W21x93,
+        W21x83,
+        W21x73,
+        W21x68,
+        W21x62,
+        W21x57,
+        W21x55,
+        W21x50,
+        W21x48,
+        W21x44,
+        W18x311,
+        W18x283,
+        W18x258,
+        W18x234,
+        W18x211,
+        W18x192,
+        W18x175,
+        W18x158,
+        W18x143,
+        W18x130,
+        W18x119,
+        W18x106,
+        W18x97,
+        W18x86,
+        W18x76,
+        W18x71,
+        W18x65,
+        W18x60,
+        W18x55,
+        W18x50,
+        W18x46,
+        W18x40,
+        W18x35,
+        W16x100,
+        W16x89,
+        W16x77,
+        W16x67,
+        W16x57,
+        W16x50,
+        W16x45,
+        W16x40,
+        W16x36,
+        W16x31,
+        W16x26,
+        W14x808,
+        W14x730,
+        W14x665,
+        W14x605,
+        W14x550,
+        W14x500,
+        W14x455,
+        W14x426,
+        W14x398,
+        W14x370,
+        W14x342,
+        W14x311,
+        W14x283,
+        W14x257,
+        W14x233,
+        W14x211,
+        W14x193,
+        W14x176,
+        W14x159,
+        W14x145,
+        W14x132,
+        W14x120,
+        W14x109,
+        W14x99,
+        W14x90,
+        W14x82,
+        W14x74,
+        W14x68,
+        W14x61,
+        W14x53,
+        W14x48,
+        W14x43,
+        W14x38,
+        W14x34,
+        W14x30,
+        W14x26,
+        W14x22,
+        W12x336,
+        W12x305,
+        W12x279,
+        W12x252,
+        W12x230,
+        W12x210,
+        W12x190,
+        W12x170,
+        W12x152,
+        W12x136,
+        W12x120,
+        W12x106,
+        W12x96,
+        W12x87,
+        W12x79,
+        W12x72,
+        W12x65,
+        W12x58,
+        W12x53,
+        W12x50,
+        W12x45,
+        W12x40,
+        W12x35,
+        W12x30,
+        W12x26,
+        W12x22,
+        W12x19,
+        W12x16,
+        W12x14,
+        W10x112,
+        W10x100,
+        W10x88,
+        W10x77,
+        W10x68,
+        W10x60,
+        W10x54,
+        W10x49,
+        W10x45,
+        W10x39,
+        W10x33,
+        W10x30,
+        W10x26,
+        W10x22,
+        W10x19,
+        W10x17,
+        W10x15,
+        W10x12,
+        W8x67,
+        W8x58,
+        W8x48,
+        W8x40,
+        W8x35,
+        W8x31,
+        W8x28,
+        W8x24,
+        W8x21,
+        W8x18,
+        W8x15,
+        W8x13,
+        W8x10,
+        W6x25,
+        W6x20,
+        W6x16,
+        W6x15,
+        W6x12,
+        W6x9,
+        W6x8_5,
+        W5x19,
+        W5x16,
+        W4x13,
+    }
+}
